обернула в функцию
