+++ index.html
@@ -231,7 +231,6 @@
   </template>
 
   <script type= "module" src = "/js/main.js"></script>
-  <script type= "module" src = "/js/similar.js"></script>
 
 </body>
 </html>
