+++ js/main.js
@@ -1,5 +1,9 @@
 import {getAds} from './data.js';
+import {renderAds} from './renderAds.js';
+
 
 // eslint-disable-next-line no-unused-vars
 const similarAds = getAds();
+// eslint-disable-next-line no-unused-vars
+const renderedAds = renderAds();
